requirements.txt
asgiref==3.3.0
Django==2.2
djangorestframework==3.10.0
gunicorn==20.0.4
pytz==2020.4
sqlparse==0.4.1
